Webhook/requirements.txt
fastapi==0.85.0
fuzzyset2==0.2.0
pydantic==1.9.2
SPARQLWrapper==2.0.0
starlette==0.20.4
uvicorn==0.18.3
gunicorn==20.1.0
requests==2.25.1
